server/requirements.txt
django
gunicorn
dj-database-url
whitenoise
psycopg2-binary
djangorestframework